% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Example_Reference_Points}
\alias{Example_Reference_Points}
\title{Habitat type point locations}
\format{
A SpatialPointsDataFrame  (package:sp) with 7 points:
\describe{
\item{Data$class:}{deciduous", "coniferous", "heath_young", "heath_old", "heath_shrub", "bare_ground", "xeric_grass", }
\item{Projection:}{"+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0". }
}
}
\source{
German Research Centre for Geosciences GFZ
}
\usage{
Example_Reference_Points
}
\description{
Spatial points of habitat types marked on Sentinel_Stack_2018
}
\details{
Example data set for habitat types that can be marked as spatial points in a known area of the input image
}
\keyword{datasets}
