% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Example_Reference_Table}
\alias{Example_Reference_Table}
\title{Habitat type spectra}
\format{
A Data.Frame with 7 observations and 54 variables:
\describe{
\item{rows:}{class.names - deciduous", "coniferous", "heath_young", "heath_old", "heath_shrub", "bare_ground", "xeric_grass", }
\item{columns:}{spectral predictors either extracted from spectral library with columns = image layers or extracted from spatial point locations. }
}
}
\source{
German Research Centre for Geosciences GFZ
}
\usage{
Example_Reference_Table
}
\description{
Habitat types (rows) and corresponding spectral predictors (columns)
}
\details{
Example data set for habitat types that can be extracted from imagery using spatial points or that is generated from a reference databases (spectral libary) with spectral predictors that correspond to the layers of the input image
}
\keyword{datasets}
