% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactive.r
\name{iplot}
\alias{iplot}
\title{Plot Habitat Types}
\usage{
iplot(x, y, HaTy, r, g, b, acc, outPath)
}
\arguments{
\item{x}{probability image (*rasterObject)}

\item{y}{RGB image (*rasterObject)}

\item{HaTy}{name of habitat type (character)}

\item{r}{red channel (integer)}

\item{g}{green channel (integer)}

\item{b}{blue channel (integer)}

\item{acc}{predictive accuracy (integer)}

\item{outPath}{file path for '.html export (character)}
}
\description{
A quick wrapper to produce an interactive raster map of habitat type probability in a web browser using leaflet
}
