% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Sentinel_Stack_2018}
\alias{Sentinel_Stack_2018}
\title{Sentinel-2A/B satellite time series stack}
\format{
A RasterBrick object  (package:raster) with 54 layers:
\describe{
\item{Wavebands:}{"blue", "green", "red", "redEdge1", "redEdge2", "redEdge3", "nearInfrared", "shortwaveInfrared1", "shortwaveInfrared2", }
\item{Dates:}{"2018-03-18", "2018-04-17", "2018-05-07", "2018-06-06", "2018-07-16", "2018-07-31", "2018-09-09", "2018-09-19", "2018-10-14", }
\item{Pixel Size:}{resampled to 10m, }
\item{Region:}{Kyritz-Ruppiner Heide, NE Germany, Brandenburg.}
}
}
\source{
Data: ESA European Space Agency

Processing: GTS2 German Research Centre for Geosciences GFZ
}
\usage{
Sentinel_Stack_2018
}
\description{
co-registrated and atmospherically corrected (reflectance) 9-waveband scenes for an entire phenological cylce in 2018
}
\details{
The area recorded is a open heathland (dominant species:Calluna vulgaris) under different management status on a former military training ground in NE Germany, trees are mainly pine, birch and oak
}
\keyword{datasets}
