% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.r
\name{plot_Results}
\alias{plot_Results}
\title{Plot Habitat Types}
\usage{
plot_Results(inPath, threshold, color = NULL)
}
\arguments{
\item{inPath}{file path (character) for results of habitat type sampling and probability mapping (same as outPath from function multi_Class_Sampling)}

\item{threshold}{probability threshold provided by the user.}

\item{color}{colors for different habitat types, have to be a color vector of length number of habitat types, there is a default colorRamp provided}
}
\value{
three plot windows
\enumerate{
\item raster map of habitat type distribution
\item pie chart of habitat type proportions
\item a raster map with delineated habitat types -> HabitatMap_final.tif
}
}
\description{
A quick wrapper to produce a habitat type map and habitat type proportions in a pie chart.
}
